split astvisitorexplainer class out over multiple smaller files
for readability and maintainability
